JOEL4: x86 segments should be 16-bit values
When loading the task register with LTR, a 16-bit segment must be used.
A bug in the macro led the LTR instruction to try to load from a 32-bit
register, which failed during assembly.  This bug was only seen when
compiling with O1 or greater, since a memory location was used with O0
which does not exhibit the same behavior.

+++ kernel/joel4/include/kstdio.h
@@ -11,7 +11,7 @@
 
 #define __KERNEL_RETURN()              __ASM("leave\nlret\n")
 
-#define SEGMENT_INDEX(_index, _ldt, _priv)  (((_index & 0x1F) << 3) | ((_ldt & 1) << 2) | (_priv & 0x3))   ///< makes a segment index
+#define SEGMENT_INDEX(_index, _ldt, _priv)  (unsigned short)(((_index & 0x1F) << 3) | ((_ldt & 1) << 2) | (_priv & 0x3))   ///< makes a segment index
 
 typedef long unsigned int size_t; // JKN - leave this for the stdlib (pdclib)
